fixing locations and other fixes
